Migrate Mutation.addAlertChannel
